SD Card Access Update, see issue #7
This should probably have been a separate issue, since the output format is unchanged as of yet.

Records are now written to a buffer.
If the new buffer size exceeds a maximum, the previous buffer is overwritten.
If the new record cannot fit into the buffer over the old record, it is dropped.

SD output file is written to in 512 byte blocks in one time slot,
then in the next time slot, the file is flushed - this keeps file system updates separate from file content updates, and prevents excessive write times.