Fixes issue #15
PID's invert flag was set by a static initialiser, which was not updated after adding extra fields into the struct.
PID is now initialised explicitly, and there is a button on the config screen to toggle the invert flag. The default is inverted, which is correct for this setup. The flag is not currently stored in eeprom.