RPM PID source configurable
RPM PID can now use rotation as it's control variable, rather than average rotation time.
This change improves resolution slightly at high rpm by reducing truncation.
This also make the RPM target control linear.
The PID can be set to use either mode using the button that formerly inverted the output. The flag is now stored in eeprom also, but may be removed if one mode proves to be superior to the other in the long run.

see issues #15 and #11