Variable pwm bits (untested)
The number of bits used for pwm can be changed. PWM is still set using a 0-255 value and is bit shifted to match the specified bits.
The idea is to keep the pwm clock at maximum and to provide a wider ranger of pwm frequencies.
As a bonus, look up tables can be higher resolution (yet to be implemented).