DAC Calibration
Compiles, untested.

Previous commit checked and debugged.
Improved documentation for some functions, improved readability with code folding.

DAC calibration implemented as a 'test type'.
DAC cal can be called with the K command followed by a non zero integer, or the 'Q' command, followed by the enum value for the calibration TEST_TYPE.
The resulting calibration can then be save with the 'E' command, also followed by a non zero integer.

Commands can now be queries if they are followed by '?'
The result could be the current value for any kind of setter, or the options available for commands that select from a list, or otherwise information about how to use the command.